MC283874 - Classic EAC Retirement Announcement
Classic Exchange Admin Center (EAC) Retirement Announcement

In MC252053 (April 2021), we announced the general availability of the new Exchange admin center (EAC). The new EAC offers a faster, more secure, and more intelligent modern admin experience than the classic EAC and is helping admins work more efficiently. With the new EAC capable of assuming this responsibility and more, today we are at the next stage of our journey, and we are announcing the retirement of the classic EAC that will be fully retired on September 1, 2022.

This will be carried out in phases as we move forward. The initial phase will be moving the features under Protection to the Microsoft 365 Defender portal.

Key Points:

Major: Retirement
Timing:
October 15, 2021 – Move Protection and Advanced Threat features to the Microsoft 365 Defender portal
September 1, 2022 – Classic EAC will be retired